Feature: shortcut completed files when resuming.
As an optimization, if the target file exists and gdown is in
resume mode, don't even try to transfer the file again.
Obviously, this only makes sense when the atomic tempfile+rename
pattern is in use, meaning only when output is to a path.

This avoids unnecessary requests. Even a ranged GET with 0 bytes
left counts against download limits, and it's just faster. The
--continue/ resume flag is handled at both the file level and the
folder level, to avoid even more requests.

This noticeably improves performance and when mirroring folders
with lots of files, when attempting to resume the transfer at a
later time. It also allows better resumption of folder downloads
that got interrupted.

Update the function docs.

Update help text for --continue to cover the new behaviour.