download: now check tempfiles more carefully.
Now that we're handling fully transferred files properly,
the tempfile finder loop should only pick files ending with
the right suffix. This avoids it finding the target file itself
then trying to update it with a (0 byte) partial transfer.

It also tries harder to avoid a possible source of file
corruption when folders contain a lot of similarly named files.